Refactor data selector
