+++ src/plugins/profiling/public/app.tsx
@@ -10,6 +10,7 @@ import React, { useState } from 'react';
 import ReactDOM from 'react-dom';
 import { AppMountParameters } from 'kibana/public';
 
+import dateMath from '@elastic/datemath';
 import {
   EuiPage,
   EuiPageBody,
@@ -17,6 +18,7 @@ import {
   EuiPageContentBody,
   EuiPageHeader,
   EuiSpacer,
+  EuiSuperDatePicker,
   EuiTabbedContent,
 } from '@elastic/eui';
 
@@ -34,9 +36,68 @@ import { PixiFlamechart } from './components/PixiFlamechart';
 
 import { Services } from './services';
 
+interface CommonlyUsedRange {
+  start: string;
+  end: string;
+  label: string;
+}
+
+const commonlyUsedRanges: CommonlyUsedRange[] = [
+  {
+    start: 'now-30m',
+    end: 'now',
+    label: 'Last 30 minutes',
+  },
+  {
+    start: 'now-1h',
+    end: 'now',
+    label: 'Last hour',
+  },
+  {
+    start: 'now-24h',
+    end: 'now',
+    label: 'Last 24 hours',
+  },
+  {
+    start: 'now-1w',
+    end: 'now',
+    label: 'Last 7 days',
+  },
+  {
+    start: 'now-30d',
+    end: 'now',
+    label: 'Last 30 days',
+  },
+];
+
+interface TimeRange {
+  start: string;
+  end: string;
+  isoStart: string;
+  isoEnd: string;
+  unixStart: number;
+  unixEnd: number;
+}
+
+function buildTimeRange(start: string, end: string): TimeRange {
+  const timeStart = dateMath.parse(start);
+  const timeEnd = dateMath.parse(end);
+  return {
+    start,
+    end,
+    isoStart: timeStart.toISOString(),
+    isoEnd: timeEnd.toISOString(),
+    unixStart: timeStart.utc().unix(),
+    unixEnd: timeEnd.utc().unix(),
+  };
+}
+
 type Props = Services;
 
 function App({ fetchTopN, fetchElasticFlamechart, fetchPixiFlamechart }: Props) {
+  const defaultTimeRange = buildTimeRange(commonlyUsedRanges[0].start, commonlyUsedRanges[0].end);
+  const [timeRange, setTimeRange] = useState(defaultTimeRange);
+
   const [index, setIndex] = useState('profiling-events-all');
   const [projectID, setProjectID] = useState(5);
   const [n, setN] = useState(100);
@@ -48,6 +109,15 @@ function App({ fetchTopN, fetchElasticFlamechart, fetchPixiFlamechart }: Props) 
 
   const [elasticFlamegraph, setElasticFlamegraph] = useState({ leaves: [] });
   const [pixiFlamegraph, setPixiFlamegraph] = useState({});
+
+  const handleTimeChange = (selectedTime: { start: string; end: string; isInvalid: boolean }) => {
+    if (selectedTime.isInvalid) {
+      return;
+    }
+
+    const tr = buildTimeRange(selectedTime.start, selectedTime.end);
+    setTimeRange(tr);
+  };
 
   const updateIndex = (idx: string) => setIndex(idx);
   const updateProjectID = (n: number) => setProjectID(n);
@@ -65,6 +135,7 @@ function App({ fetchTopN, fetchElasticFlamechart, fetchPixiFlamechart }: Props) 
               index={index}
               projectID={projectID}
               n={n}
+              timeRange={timeRange}
               fetchTopN={fetchTopN}
               setTopN={setTopN}
             />
@@ -85,6 +156,7 @@ function App({ fetchTopN, fetchElasticFlamechart, fetchPixiFlamechart }: Props) 
               index={index}
               projectID={projectID}
               n={n}
+              timeRange={timeRange}
               getter={fetchElasticFlamechart}
               setter={setElasticFlamegraph}
             />
@@ -104,6 +176,7 @@ function App({ fetchTopN, fetchElasticFlamechart, fetchPixiFlamechart }: Props) 
               index={index}
               projectID={projectID}
               n={n}
+              timeRange={timeRange}
               getter={fetchPixiFlamechart}
               setter={setPixiFlamegraph}
             />
@@ -121,6 +194,13 @@ function App({ fetchTopN, fetchElasticFlamechart, fetchPixiFlamechart }: Props) 
           paddingSize="s"
           pageTitle="Continuous Profiling"
           rightSideItems={[
+            <EuiSuperDatePicker
+              start={timeRange.start}
+              end={timeRange.end}
+              isPaused={true}
+              onTimeChange={handleTimeChange}
+              commonlyUsedRanges={commonlyUsedRanges}
+            />,
             <SettingsFlyout
               defaultIndex={index}
               updateIndex={updateIndex}
