+++ src/plugins/profiling/public/components/stacktrace-nav.tsx
@@ -8,11 +8,11 @@
 
 import React, { useEffect, useState } from 'react';
 
-import { EuiButtonGroup, EuiSpacer } from '@elastic/eui';
+import { EuiButtonGroup } from '@elastic/eui';
 
 import { getTopN, groupSamplesByCategory } from '../../common';
 
-export const StackTraceNavigation = ({ index, projectID, n, fetchTopN, setTopN }) => {
+export const StackTraceNavigation = ({ index, projectID, n, timeRange, fetchTopN, setTopN }) => {
   const topnButtonGroupPrefix = 'topnButtonGroup';
 
   const topnButtons = [
@@ -52,63 +52,22 @@ export const StackTraceNavigation = ({ index, projectID, n, fetchTopN, setTopN }
     setToggleTopNSelected(optionId);
   };
 
-  const dateButtonGroupPrefix = 'dateButtonGroup';
-
-  const dateButtons = [
-    {
-      id: `${dateButtonGroupPrefix}__0`,
-      label: '30m',
-      value: '1800',
-    },
-    {
-      id: `${dateButtonGroupPrefix}__1`,
-      label: '1h',
-      value: '3600',
-    },
-    {
-      id: `${dateButtonGroupPrefix}__2`,
-      label: '24h',
-      value: '86400',
-    },
-    {
-      id: `${dateButtonGroupPrefix}__3`,
-      label: '7d',
-      value: '604800',
-    },
-    {
-      id: `${dateButtonGroupPrefix}__4`,
-      label: '30d',
-      value: '2592000',
-    },
-  ];
-
-  const [toggleDateSelected, setToggleDateSelected] = useState(`${dateButtonGroupPrefix}__0`);
-
-  const onDateChange = (optionId: string) => {
-    if (optionId === toggleDateSelected) {
-      return;
-    }
-    setToggleDateSelected(optionId);
-  };
-
   useEffect(() => {
     const topnValue = topnButtons.filter((button) => {
       return button.id === toggleTopNSelected;
     });
 
-    const dateValue = dateButtons.filter((button) => {
-      return button.id === toggleDateSelected;
-    });
-
     console.log(new Date().toISOString(), 'started payload retrieval');
-    fetchTopN(index, projectID, topnValue[0].value, dateValue[0].value, n).then((response) => {
-      console.log(new Date().toISOString(), 'finished payload retrieval');
-      const samples = getTopN(response);
-      const series = groupSamplesByCategory(samples);
-      setTopN({ samples, series });
-      console.log(new Date().toISOString(), 'updated local state');
-    });
-  }, [projectID, toggleTopNSelected, toggleDateSelected]);
+    fetchTopN(topnValue[0].value, index, projectID, timeRange.unixStart, timeRange.unixEnd, n).then(
+      (response) => {
+        console.log(new Date().toISOString(), 'finished payload retrieval');
+        const samples = getTopN(response);
+        const series = groupSamplesByCategory(samples);
+        setTopN({ samples, series });
+        console.log(new Date().toISOString(), 'updated local state');
+      }
+    );
+  }, [index, projectID, n, timeRange, toggleTopNSelected]);
 
   return (
     <div>
@@ -117,13 +76,6 @@ export const StackTraceNavigation = ({ index, projectID, n, fetchTopN, setTopN }
         options={topnButtons}
         idSelected={toggleTopNSelected}
         onChange={(id) => onTopNChange(id)}
-      />
-      <EuiSpacer size="s" />
-      <EuiButtonGroup
-        legend="This is a basic group"
-        options={dateButtons}
-        idSelected={toggleDateSelected}
-        onChange={(id) => onDateChange(id)}
       />
     </div>
   );
